+++ src/Sulu/Bundle/WebsiteBundle/Tests/Unit/Sulu/Bundle/WebsiteBundle/Routing/PortalLoaderTest.php
@@ -12,6 +12,7 @@
 namespace Sulu\Bundle\WebsiteBundle\Routing;
 
 use Prophecy\Argument;
+use Sulu\Component\Webspace\Analyzer\RequestAnalyzerInterface;
 use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
 use Sulu\Component\Webspace\Portal;
 use Sulu\Component\Webspace\PortalInformation;
@@ -89,5 +90,49 @@ class PortalLoaderTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('/de/route2', $routeCollection->get('sulu.io/de.route2')->getPath());
         $this->assertEquals('/example/route1', $routeCollection->get('sulu.com.route1')->getPath());
         $this->assertEquals('/route2', $routeCollection->get('sulu.com.route2')->getPath());
+    }
+
+    public function testLoadPartial()
+    {
+        $importedRouteCollection = new RouteCollection();
+        $importedRouteCollection->add('route', new Route('/route'));
+
+        $portal = new Portal();
+        $portal->setKey('sulu_lo');
+
+        $portalInformations = [
+            new PortalInformation(null, null, $portal, null, 'sulu.io/de'),
+            new PortalInformation(
+                RequestAnalyzerInterface::MATCH_TYPE_PARTIAL,
+                null,
+                $portal,
+                null,
+                'sulu.io',
+                null,
+                'sulu.io/de'
+            ),
+        ];
+
+        $this->loaderResolver->resolve(Argument::any(), Argument::any())->willReturn($this->loader->reveal());
+        $this->loader->load(Argument::any(), Argument::any())->willReturn($importedRouteCollection);
+        $this->webspaceManager->getPortalInformations(Argument::any())->willReturn($portalInformations);
+
+        $routeCollection = $this->portalLoader->load('', 'portal');
+
+        $this->assertCount(2, $routeCollection);
+
+        $routes = $routeCollection->getIterator();
+        $this->assertArrayHasKey('sulu.io.route', $routes);
+        $this->assertArrayHasKey('sulu.io/de.route', $routes);
+        $this->assertEquals('/route', $routeCollection->get('sulu.io.route')->getPath());
+        $this->assertEquals('/de/route', $routeCollection->get('sulu.io/de.route')->getPath());
+        $this->assertEquals(
+            [
+                '_controller' => 'SuluWebsiteBundle:Redirect:redirectToRoute',
+                'route' => 'sulu.io/de.route',
+                'permanent' => true,
+            ],
+            $routeCollection->get('sulu.io.route')->getDefaults()
+        );
     }
 }